Сделан доступ для удаления комментариев

+++ src/main/java/ru/skypro/homework/service/impl/AdsServiceImpl.java
@@ -2,6 +2,7 @@ package ru.skypro.homework.service.impl;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import ru.skypro.homework.controller.UserController;
 import ru.skypro.homework.dto.AdsComment;
@@ -11,11 +12,13 @@ import ru.skypro.homework.dto.FullAdsDto;
 import ru.skypro.homework.entity.Ads;
 import ru.skypro.homework.entity.Comment;
 import ru.skypro.homework.entity.UserProfile;
+import ru.skypro.homework.exception.AccessDeniedException;
 import ru.skypro.homework.exception.CommentNotFoundException;
 import ru.skypro.homework.mapper.AdsMapper;
 import ru.skypro.homework.mapper.CommentMapper;
 import ru.skypro.homework.repository.AdsRepository;
 import ru.skypro.homework.repository.CommentRepository;
+import ru.skypro.homework.repository.UserProfileRepository;
 import ru.skypro.homework.service.AdsService;
 
 import java.util.Collection;
@@ -28,10 +31,12 @@ public class AdsServiceImpl implements AdsService {
 
     private final CommentRepository commentRepository;
     private final AdsRepository adsRepository;
+    private final UserProfileRepository userProfileRepository;
 
-    public AdsServiceImpl(CommentRepository commentRepository, AdsRepository adsRepository) {
+    public AdsServiceImpl(CommentRepository commentRepository, AdsRepository adsRepository, UserProfileRepository userProfileRepository) {
         this.commentRepository = commentRepository;
         this.adsRepository = adsRepository;
+        this.userProfileRepository = userProfileRepository;
     }
 
     @Override
@@ -78,8 +83,16 @@ public class AdsServiceImpl implements AdsService {
     }
 
     @Override
-    public void deleteComment(Long adsId, Long commentId) {
-        commentRepository.deleteByAdsIdAndId(adsId, commentId);
+    public void deleteComment(Long adsId, Long commentId, Authentication authentication) {
+
+        Long userIdFromComments = commentRepository.getUserProfileId(adsId, commentId);
+        Long userIdFromUserProfiles = userProfileRepository.getUserProfileId(authentication.getName());
+
+        if (userIdFromComments == userIdFromUserProfiles ||
+            authentication.getAuthorities().stream().anyMatch(a -> a.getAuthority().equals("ADMIN"))) {
+            commentRepository.deleteByAdsIdAndId(adsId, commentId);
+        }
+        throw new AccessDeniedException();
     }
 
     @Override
